move login to user module
